Output separate YAML files for cisagov conversion
Instead of printing one large YAML file to stdout we instead group the
read software entires by the uppercased first character of the vendor
name and output a file for each of these groups. All non-alphabet
vendor entries are grouped in the "0-9" group manually. When testing
the only entries that were non-alphabet were numeric. This will generate
files that are small enough to be edited using the GitHub web UI.