Use a top level key to store data ownership
Instead of creating a `reporter` key for every entry in the converted
Markdown we instead establish data ownership at in the YAML file as a
whole. This will ensure that all data in a given YAML file is attributed
to a specific source as was intended by the `reporter` key.