Sort the list of processed rows before grouping
The itertools.groupby() function only correctly functions if the list
it is grouping is pre-sorted in the same manner as the grouping key. If
the list is not similarly sorted it will result in lost list items. We
also switch to using `.lower()` instead of `.upper()` to be consistent
with other sorting in this project.